% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs2015_county_data_info.R
\docType{data}
\name{acs2015_county_data}
\alias{acs2015_county_data}
\title{acs2015_county_data includes demographic and economic census data for every county or county equivalent in the United States}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 37 columns.}
\source{
\href{https://www.kaggle.com/muonneutrino/us-census-demographic-data}{Kaggle}
}
\usage{
acs2015_county_data
}
\description{
acs2015_county_data includes demographic and economic census data for every county or county equivalent in the United States
}
\keyword{datasets}
