% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS_Honeypot_marx_geo_info.R
\docType{data}
\name{AWS_Honeypot_marx_geo}
\alias{AWS_Honeypot_marx_geo}
\title{AWS_Honeypot_marx_geo contains AWS honeypot data from 9:53 PM on 3 March 2013 to 5:55 AM on 8 September 2013}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451581 rows and 16 columns.}
\source{
\href{https://www.kaggle.com/casimian2000/aws-honeypot-attack-data}{Kaggle}
}
\usage{
AWS_Honeypot_marx_geo
}
\description{
AWS_Honeypot_marx_geo contains AWS honeypot data from 9:53 PM on 3 March 2013 to 5:55 AM on 8 September 2013
}
\keyword{datasets}
