% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_coded_info.R
\docType{data}
\name{food_coded}
\alias{food_coded}
\title{food_coded contains data on college students and their food choices}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125 rows and 61 columns.}
\source{
\href{https://www.kaggle.com/borapajo/food-choices}{Kaggle}
}
\usage{
food_coded
}
\description{
food_coded contains data on college students and their food choices
}
\keyword{datasets}
