% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_wrapper.R
\name{ggplot_wrapper}
\alias{ggplot_wrapper}
\title{ggplot_wrapper graphs a function using ggplot2}
\usage{
ggplot_wrapper(dat, x, y, color)
}
\arguments{
\item{dat}{the data}

\item{x}{the x value for ggplot2}

\item{y}{the y value for ggplot2}

\item{color}{optional color value for ggplot2}
}
\description{
ggplot_wrapper graphs a function using ggplot2
}
\examples{
\dontrun{
 ggplot_wrapper(data, data$x, data$y, "red")
}
}
