% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-xax-grapes.R
\name{\%xax\%}
\alias{\%xax\%}
\title{\%xax\% calculates the transverse of vector x multiplied by the inverse of A multiplied}
\usage{
a \%xax\% x
}
\arguments{
\item{a}{numerical vector}

\item{x}{numerical vector}
}
\value{
1 x 1 matrix containing the operation result
}
\description{
\%xax\% calculates the transverse of vector x multiplied by the inverse of A multiplied
}
\examples{
\dontrun{
 a \%xax\% x
}
}
