% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{mle calculates the maximum likelihood estimate for the gamma distribution}
\usage{
mle(x)
}
\arguments{
\item{x}{the numerical vector to optimize over}
}
\value{
the value of the distribution that gives the highest value
}
\description{
mle calculates the maximum likelihood estimate for the gamma distribution
}
\examples{
\dontrun{
 mle(x)
}
}
