% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myapply.R
\name{myapply}
\alias{myapply}
\title{myapply is a modified version of the apply function, but it only works on margin = 1 and 2}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{the array to apply the function over}

\item{MARGIN}{1 for rows and 2 for columns}

\item{FUN}{the function to apply over}
}
\value{
the modified array
}
\description{
myapply is a modified version of the apply function, but it only works on margin = 1 and 2
}
\examples{
\dontrun{
 myapply(x, 1, mean)
}
}
