% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{opt}
\alias{opt}
\title{This function is used to estimate where the maximum is given a function}
\usage{
opt(x, f, interval)
}
\arguments{
\item{x}{the entries}

\item{f}{the function to optimize}

\item{interval}{the interval to optimize over}
}
\value{
filtered data
}
\description{
This function is used to estimate where the maximum is given a function
}
\examples{
\dontrun{
 opt(x, f, c(-1,1))
}
}
