% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{slice filters a specific column so all entries match a certain value}
\usage{
slice(dat, col, filtWith)
}
\arguments{
\item{dat}{the data}

\item{col}{the column to filter}

\item{filtWith}{what the value should equal}
}
\value{
filtered data
}
\description{
slice filters a specific column so all entries match a certain value
}
\examples{
\dontrun{
 slice(data, data$height, 6)
}
}
