% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize calculates (x - mean(x))/sd(x) for each column in a given matrix}
\usage{
standardize(a)
}
\arguments{
\item{a}{the matrix}
}
\value{
the standardized matrix
}
\description{
standardize calculates (x - mean(x))/sd(x) for each column in a given matrix
}
\examples{
\dontrun{
 standardize(x)
}
}
