% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summFunc.R
\name{summFunc}
\alias{summFunc}
\title{summFunc returns the mean, variance, and standard deviation of a numerical matrix in a list}
\usage{
summFunc(x)
}
\arguments{
\item{x}{the matrix}
}
\value{
the mean, variance, and standard deviation of x in a list
}
\description{
summFunc returns the mean, variance, and standard deviation of a numerical matrix in a list
}
\examples{
\dontrun{
 summFunc(x)
}
}
