% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedSumm.R
\name{weightedSumm}
\alias{weightedSumm}
\title{weightedSumm returns the weighted mean, variance, and standard deviation of a numerical matrix in a list}
\usage{
weightedSumm(x, p)
}
\arguments{
\item{x}{the matrix}
}
\value{
the weighted mean, variance, and standard deviation of x in a list
}
\description{
weightedSumm returns the weighted mean, variance, and standard deviation of a numerical matrix in a list
}
\examples{
\dontrun{
 weightedSumm(x)
}
}
