% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xax.R
\name{xax}
\alias{xax}
\title{xax calculates the transverse of matrix x multiplied by the inverse of A multiplied}
\usage{
xax(a, x)
}
\arguments{
\item{a}{numerical vector}

\item{x}{numerical vector}
}
\value{
1 x 1 matrix containing the operation result
}
\description{
xax calculates the transverse of matrix x multiplied by the inverse of A multiplied
}
\examples{
\dontrun{
 xax(a,x)
}
}
